Fix: 태그별 뮤멘트 카드 그림자 가려지는 문제 및 간격 조정#85
