+++ MUMENT/MUMENT/Sources/Scenes/Home/Main/MumentsByTagCVC.swift
@@ -20,44 +20,7 @@ protocol MumentsByTagCVCDelegate : AnyObject{
 class MumentsByTagCVC: UICollectionViewCell {
     
     // MARK: - Properties
-    lazy var titleSectionStackView = UIStackView(arrangedSubviews: [titleIconImage, titleAndArtistLabel]).then{
-        $0.axis = .horizontal
-        $0.spacing = 5
-        $0.alignment = .center
-//        $0.alignment = .top
-    }
-    
-    private let titleIconImage = UIImageView().then{
-        $0.image = UIImage(named: "mumentMusicnote")
-    }
-    
-    private let titleAndArtistLabel = UILabel().then{
-        $0.textColor = .mBlack2
-        $0.font = .mumentB5B13
-    }
-    
-    private let contentsLabel = UILabel().then{
-        $0.textColor = .mGray1
-        $0.font = .mumentB6M13
-        $0.lineBreakMode = .byTruncatingTail
-        $0.numberOfLines = 5
-    }
-    
-    let separatorView = UIView().then{
-        $0.backgroundColor = .mGray4
-    }
-    
-    lazy var writerInfoStackView = UIStackView(arrangedSubviews: [profileImage, writerNameLabel]).then{
-        $0.axis = .horizontal
-        $0.spacing = 6
-    }
-    private var profileImage = UIImageView().then{
-        $0.makeRounded(cornerRadius: 9.5)
-    }
-    private let writerNameLabel = UILabel().then{
-        $0.textColor = .mGray1
-        $0.font = .mumentC1R12
-    }
+    let mumentsByTagCardView = MumentsByTagCardView()
     
     // MARK: - Initialization
     override init(frame: CGRect) {
@@ -72,71 +35,34 @@ class MumentsByTagCVC: UICollectionViewCell {
     }
     
     //MARK: - Functions
-    func setData(_ cellData: MumentsByTagModel){
-        titleAndArtistLabel.text = "\(cellData.title) - \(cellData.artist)"
-        contentsLabel.text = cellData.contents
-        profileImage.image = cellData.profileImage
-        writerNameLabel.text = cellData.writerName
+    func setData(_ cellData: MumentsByTagResponseModel.MumentList){
+        mumentsByTagCardView.setData(cellData)
     }
     
-    func setData(_ cellData: MumentsByTagResponseModel.MumentList){
-        titleAndArtistLabel.text = "\(cellData.music.name) - \(cellData.music.artist)"
-        contentsLabel.text = cellData.content
-        profileImage.setImageUrl(cellData.user.image ?? "")
-        writerNameLabel.text = cellData.user.name
+    // Test Code
+    func setData(_ cellData: MumentsByTagModel){
+        mumentsByTagCardView.setData(cellData)
     }
+    
 }
 
 // MARK: - UI
 extension MumentsByTagCVC {
-    
     private func setUI(){
-        self.makeRounded(cornerRadius: 11)
-        self.backgroundColor = .mWhite
-        self.addShadow(offset: CGSize(width: 0, height: -2),opacity: 0.2,radius: 4.0)
+        
     }
     
     private func setLayout() {
+        self.addSubviews([mumentsByTagCardView])
         
-        self.addSubviews([writerInfoStackView,separatorView, titleSectionStackView,contentsLabel])
-      
-        writerInfoStackView.snp.makeConstraints{
-//            $0.top.equalTo(separatorView.snp.bottom).offset(10)
-            $0.left.equalTo(self.safeAreaLayoutGuide).offset(13)
-            $0.right.equalTo(self.safeAreaLayoutGuide).inset(13)
-            $0.bottom.equalTo(self.safeAreaLayoutGuide).inset(13)
+        mumentsByTagCardView.snp.makeConstraints{
+//            $0.edges.equalTo(self.safeAreaLayoutGuide)
+            $0.left.equalTo(self.safeAreaLayoutGuide).offset(5)
+            $0.right.equalTo(self.safeAreaLayoutGuide).offset(5)
+            $0.bottom.equalTo(self.safeAreaLayoutGuide)
+            $0.top.equalTo(self.safeAreaLayoutGuide)
         }
-        
-        separatorView.snp.makeConstraints{
-//            $0.top.equalTo(writerInfoStackView.snp.top).offset(10)
-            $0.left.equalTo(self.safeAreaLayoutGuide).offset(13)
-            $0.right.equalTo(self.safeAreaLayoutGuide).inset(13)
-            $0.bottom.equalToSuperview().inset(50)
-//            $0.top.equalTo(contentsLabel.snp.bottom).offset(10)
-//            $0.bottom.equalTo(writerInfoStackView.snp.top).offset(30)
-            $0.height.equalTo(1)
-        }
-        
-        titleSectionStackView.snp.makeConstraints{
-            $0.top.equalTo(self.safeAreaLayoutGuide).offset(19)
-            $0.left.equalTo(self.safeAreaLayoutGuide).offset(13)
-            $0.right.equalTo(self.safeAreaLayoutGuide).inset(13)
-        }
-        
-        contentsLabel.snp.makeConstraints{
-            $0.top.equalTo(titleSectionStackView.snp.bottom).offset(10)
-            $0.leading.equalTo(self.safeAreaLayoutGuide.snp.leading).offset(13)
-            $0.trailing.equalTo(self.safeAreaLayoutGuide.snp.trailing).inset(13)
-        }
-        
-        titleIconImage.snp.makeConstraints{
-            $0.height.width.equalTo(12)
-        }
-        
-        profileImage.snp.makeConstraints{
-            $0.height.width.equalTo(24)
-        }
-        
     }
+    
 }
